Decouple FrequencyData from the analysis feature
The applet renders FrequencyData (fed by the daemon's SSE bands) but never
computes spectra — yet it enabled shared's `analysis` feature, pulling in
spectrum-analyzer just for that plain data type. Move FrequencyData into an
ungated audio::types module and keep the AudioAnalyzer gated behind
`analysis`. The applet drops the feature and no longer builds
spectrum-analyzer; the daemon (which owns the analyzer) is unchanged.
